step-2: разделил один большой action.js на пачку мелких
