+++ src/components/App/App.js
@@ -6,7 +6,8 @@ import style from "./App.module.css";
 import Modal from "../Modal/Modal";
 import IngredientDetails from "../IngredientDetails/IngredientDetails";
 import OrderDetails from "../OrderDetails/OrderDetails";
-import { getIngredients, SET_CURRENT_INGREDIENT } from "../../services/actions";
+import { getIngredients } from "../../services/actions/getIngredients";
+import { SET_CURRENT_INGREDIENT } from "../../services/actions/ingredientInfo";
 import { useDispatch, useSelector } from "react-redux";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
