+++ src/services/actions/getIngredients.js
@@ -1,0 +1,27 @@
+// Получение списка ингредиентов от API. Используется в компоненте BurgerConstructor.
+import { recieveIngredients } from "../../utils/api";
+
+export const GET_INGREDIENTS_REQUEST = "GET_INGREDIENTS_REQUEST";
+export const GET_INGREDIENTS_SUCCESS = "GET_INGREDIENTS_SUCCESS";
+export const GET_INGREDIENTS_FAILED = "GET_INGREDIENTS_FAILED";
+
+export const getIngredients = () => {
+    return (dispatch) => {
+        dispatch({
+            type: GET_INGREDIENTS_REQUEST,
+        });
+        recieveIngredients()
+            .then(({ data }) => {
+                dispatch({
+                    type: GET_INGREDIENTS_SUCCESS,
+                    data: data,
+                });
+            })
+            .catch((error) => {
+                dispatch({
+                    type: GET_INGREDIENTS_FAILED,
+                });
+                console.error("Ошибка при получении данных:", error);
+            });
+    };
+};
