+++ src/services/actions/ingredientInfo.js
@@ -1,0 +1,1 @@
+export const SET_CURRENT_INGREDIENT = "SET_CURRENT_INGREDIENT";
