+++ src/services/reducers/ingredients.js
@@ -2,7 +2,7 @@ import {
     GET_INGREDIENTS_FAILED,
     GET_INGREDIENTS_REQUEST,
     GET_INGREDIENTS_SUCCESS,
-} from "../actions";
+} from "../actions/getIngredients";
 
 const initialState = {
     ingredients: [],
